+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## X.X.X 2022-XX-XX
 ### PaymentSheet
-* [Changed] The "save this card" checkbox in PaymentSheet is now unchecked by default.
+* [Changed] The "save this card" checkbox in PaymentSheet is now unchecked by default in non-US countries.
 * [Fixed] Fixes issue that could cause symbol name collisions when using Objective-C
 * [Fixed] Fixes potential crash when using PaymentSheet with SwiftUI
 * `Stripe` now requires `StripeApplePay`. See `MIGRATING.md` for more info.